feat[ADM-4]: added useBanUserForm
