+++ src/features/banUser/ui/BanUserDialog.tsx
@@ -1,35 +1,7 @@
-import { Controller, SubmitHandler, useForm } from 'react-hook-form'
-
 import { useTranslation } from '@/shared/hooks/useTranslation'
-import { zodResolver } from '@hookform/resolvers/zod'
-import { Dialog, Select, TextArea } from '@nazar-pryt/inctagram-ui-kit'
-import { z } from 'zod'
+import { Dialog } from '@nazar-pryt/inctagram-ui-kit'
 
-import { BanOptions, useBanOptions } from '../hook/useBanOptions'
-
-const maxTextAreaValueLength = 10
-
-const banUserFormSchema = z.object({
-  reasonSelectValue: z.nativeEnum(BanOptions),
-  textAreaValue: z
-    .string()
-    .trim()
-    .max(
-      maxTextAreaValueLength,
-      `Ban-reason message cant be more then ${maxTextAreaValueLength} symbols!!`
-    ),
-})
-
-export type BanUserFormSchemaType = z.infer<typeof banUserFormSchema>
-
-type SchemaType = {
-  [key in BanOptions]: keyof BanUserFormSchemaType
-}
-const Schema: SchemaType = {
-  ['advertising-placement']: 'reasonSelectValue',
-  ['another-reason']: 'textAreaValue',
-  ['bad-behavior']: 'reasonSelectValue',
-}
+import { useBanUserForm } from '../hook/useBanUserForm'
 
 type BanUserDialogType = {
   banDialog: boolean
@@ -46,29 +18,8 @@ export const BanUserDialog = ({
   userName,
 }: BanUserDialogType) => {
   const { t } = useTranslation()
-  const {
-    control,
-    formState: { errors },
-    handleSubmit,
-    setError,
-    watch,
-  } = useForm<BanUserFormSchemaType>({
-    defaultValues: { reasonSelectValue: undefined, textAreaValue: '' }, // todo fix possibility to pass null as default value
-    resolver: zodResolver(banUserFormSchema),
-  })
 
-  const onSubmit: SubmitHandler<BanUserFormSchemaType> = data => {
-    const sendingKeyFromData = Schema[data.reasonSelectValue]
-
-    if (data[sendingKeyFromData].length) {
-      handleBanUser(data[sendingKeyFromData])
-    } else {
-      setError('textAreaValue', { message: 'Ban-reason message is required' })
-    }
-  }
-  const banOptions = useBanOptions()
-
-  const showTextArea = watch('reasonSelectValue') === BanOptions.anotherReason
+  const { renderBanUserForm, submitBanUserForm } = useBanUserForm({ handleBanUser, userName })
 
   return (
     <Dialog
@@ -78,45 +29,11 @@ export const BanUserDialog = ({
       invertButtons
       onCancelButtonClick={handleCloseBanDialog}
       onClose={handleCloseBanDialog}
-      onConfirmButtonClick={handleSubmit(onSubmit)}
+      onConfirmButtonClick={submitBanUserForm}
       open={banDialog}
       title={t.user_list_popover.ban}
     >
-      <form onSubmit={handleSubmit(onSubmit)}>
-        <p>
-          {t.user_list_popover.ban_question}, {userName}?
-        </p>
-        <Controller
-          control={control}
-          name={'reasonSelectValue'}
-          render={({ field: { onChange, value } }) => (
-            <Select
-              onChange={onChange}
-              options={banOptions}
-              placeholder={t.user_list_popover.ban_reason}
-              portal={false}
-              value={value}
-            />
-          )}
-        />
-
-        {showTextArea && (
-          <Controller
-            control={control}
-            name={'textAreaValue'}
-            render={({ field: { onChange, value } }) => (
-              <TextArea
-                autoFocus
-                error={errors.textAreaValue?.message}
-                maxLength={maxTextAreaValueLength}
-                onChange={onChange}
-                style={{ backgroundColor: 'gray' }}
-                value={value}
-              />
-            )}
-          />
-        )}
-      </form>
+      {renderBanUserForm()}
     </Dialog>
   )
 }
